Do not separate db writes
