+++ massa-execution-worker/Cargo.toml
@@ -37,7 +37,6 @@ massa_ledger_exports = { path = "../massa-ledger-exports" }
 massa_pos_exports = { path = "../massa-pos-exports" }
 massa_final_state = { path = "../massa-final-state" }
 massa_versioning = { path = "../massa-versioning" }
-massa_db_exports = { path = "../massa-db-exports" }
 massa_db_worker = { path = "../massa-db-worker", optional = true }
 massa_wallet = { path = "../massa-wallet" }
